Repurpose node-osh-path for simplicity
No longer wrap superagent to make requests "easier." This is now a library for
converting URIs to POJOs and back. That's it.

This version is unnecessarily complex. Currently, it parses AND validates query
parameters, which raises several tricky questions such as: What to do when a
query parameter is invalid? Include it in the URI/POJO? Ignore it, and let the
user know the route matched but got an invalid parameter? Error? These
decisions are not for this library; they should be delegated to the page logic.
Future versions will simply parse/build query strings without assumption.